fix: add RandomEventState to all scenarios and fix cooldown overflow
Add RandomEventState and randomEventFrequency to the 5 new scenario
files that were created before the RandomEventModule merge. Fix integer
overflow in cooldown check where (year - Integer.MIN_VALUE) wrapped
negative, blocking all events from firing.

+++ src/main/java/org/flossware/civilization/module/RandomEventModule.java
@@ -42,7 +42,7 @@ public final class RandomEventModule implements SimulationModule {
         }
 
         int currentYear = state.year();
-        if (currentYear - state.randomEvent().lastEventYear() < COOLDOWN_YEARS) {
+        if ((long) currentYear - (long) state.randomEvent().lastEventYear() < COOLDOWN_YEARS) {
             return new ModuleResult<>(state, List.of());
         }
